fix vhea field sizes
The version number has 32 bits. The of advanceHeightMax has been changed
to unsigned, which corresponds to the Microsoft documentation. Apple
docs say that it is signed, but that does not make sense.

+++ packages/fontkit/src/tables/vhea.ts
@@ -46,11 +46,11 @@ export namespace vheaTable {
 }
 
 const vheaFields = {
-	version: r.uint16, // Version number of the Vertical Header Table
+	version: r.uint32, // Version number of the Vertical Header Table
 	ascent: r.int16, // The vertical typographic ascender for this font
 	descent: r.int16, // The vertical typographic descender for this font
 	lineGap: r.int16, // The vertical typographic line gap for this font
-	advanceHeightMax: r.int16, // The maximum advance height measurement found in the font
+	advanceHeightMax: r.uint16, // The maximum advance height measurement found in the font
 	minTopSideBearing: r.int16, // The minimum top side bearing measurement found in the font
 	minBottomSideBearing: r.int16, // The minimum bottom side bearing measurement found in the font
 	yMaxExtent: r.int16,
